Update the showdb.c utility so that it works with databases larger than 4GiB. (CVS 6656)
FossilOrigin-Name: 9675b6223645fe01b202cbefae63f3209a50efcc

+++ tool/showdb.c
@@ -25,7 +25,7 @@ static print_page(int iPg){
   int i, j;
   aData = malloc(pagesize);
   if( aData==0 ) out_of_memory();
-  lseek(db, (iPg-1)*pagesize, SEEK_SET);
+  lseek(db, (iPg-1)*(long long int)pagesize, SEEK_SET);
   read(db, aData, pagesize);
   fprintf(stdout, "Page %d:\n", iPg);
   for(i=0; i<pagesize; i += perLine){
